Add support for slots and events.
These are parsed independently of the hybrids definition in case they are defined elsewhere in the module. However, they rely on the custom element declaration to be in the context object. This may create some unfortunate documentation misses when dispatch or slot are used before define in the the custom element module.